pyproject.toml
[tool.poetry]
name = "humidity-sensor"
version = "0.1.0"
description = ""
authors = ["tomaszbk-pc <zubik.tomas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.24.0.post1"
paho-mqtt = "^1.6.1"
loguru = "^0.7.2"
pymongo = "^4.6.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
